+++ test/passing/tests/comments_around_disabled.ml
@@ -5,4 +5,13 @@ let () =
   ()
 [@@@ocamlformat "enable"]
 
+[@@@ocamlformat "disable"]
+ (* x *)
+ (* y *)
+let x =
+  x
+(* z *)
+
+[@@@ocamlformat "enable"]
+
 (* cmts *)