+++ CHANGES.md
@@ -8,6 +8,7 @@
 
 - Avoid adding breaks inside `~label:(fun` and base the indentation on the label. (#2271, @Julow)
 - Fix non-stabilizing comments attached to private/virtual/mutable keywords (#2272, @gpetiot)
+- Fix formatting of comments in "disable" chunks (#2279, @gpetiot)
 
 ### Changes